ref: move MyInfoService into public/services
