+++ src/public/modules/forms/admin/controllers/activate-form-modal.client.controller.js
@@ -1,5 +1,5 @@
 const dedent = require('dedent-js')
-const MyInfoService = require('../../services/myinfo.service')
+const MyInfoService = require('../../../../services/MyInfoService')
 
 angular
   .module('forms')
